+++ mobile/android/android-components/components/browser/engine-gecko-nightly/src/test/java/mozilla/components/browser/engine/gecko/prompt/GeckoPromptDelegateTest.kt
@@ -1084,6 +1084,98 @@ class GeckoPromptDelegateTest {
     }
 
     @Test
+    fun `onRepostConfirmPrompt must provide a Repost PromptRequest`() {
+        val mockSession = GeckoEngineSession(runtime)
+        var request: PromptRequest.Repost = mock()
+        var onPositiveButtonWasCalled = false
+        var onNegativeButtonWasCalled = false
+
+        mockSession.register(object : EngineSession.Observer {
+            override fun onPromptRequest(promptRequest: PromptRequest) {
+                request = promptRequest as PromptRequest.Repost
+            }
+        })
+
+        val promptDelegate = GeckoPromptDelegate(mockSession)
+
+        var geckoResult = promptDelegate.onRepostConfirmPrompt(mock(), GeckoRepostPrompt())
+        geckoResult!!.accept {
+            onPositiveButtonWasCalled = true
+        }
+        request.onConfirm()
+        assertTrue(onPositiveButtonWasCalled)
+
+        geckoResult = promptDelegate.onRepostConfirmPrompt(mock(), GeckoRepostPrompt())
+        geckoResult!!.accept {
+            onNegativeButtonWasCalled = true
+        }
+        request.onDismiss()
+        assertTrue(onNegativeButtonWasCalled)
+    }
+
+    @Test
+    fun `onRepostConfirmPrompt will not be able to complete multiple times`() {
+        val mockSession = GeckoEngineSession(runtime)
+        var request: PromptRequest.Repost = mock()
+
+        mockSession.register(object : EngineSession.Observer {
+            override fun onPromptRequest(promptRequest: PromptRequest) {
+                request = promptRequest as PromptRequest.Repost
+            }
+        })
+
+        val promptDelegate = GeckoPromptDelegate(mockSession)
+
+        var prompt = mock<GeckoRepostPrompt>()
+        promptDelegate.onRepostConfirmPrompt(mock(), prompt)
+        doReturn(false).`when`(prompt).isComplete
+        request.onConfirm()
+        verify(prompt).confirm(any())
+
+        prompt = mock()
+        promptDelegate.onRepostConfirmPrompt(mock(), prompt)
+        doReturn(true).`when`(prompt).isComplete
+        request.onConfirm()
+        verify(prompt, never()).confirm(any())
+
+        prompt = mock()
+        promptDelegate.onRepostConfirmPrompt(mock(), prompt)
+        doReturn(false).`when`(prompt).isComplete
+        request.onDismiss()
+        verify(prompt).confirm(any())
+
+        prompt = mock()
+        promptDelegate.onRepostConfirmPrompt(mock(), prompt)
+        doReturn(true).`when`(prompt).isComplete
+        request.onDismiss()
+        verify(prompt, never()).confirm(any())
+    }
+
+    @Test
+    fun `onRepostConfirmPrompt will inform listeners when it is being dismissed`() {
+        val mockSession = GeckoEngineSession(runtime)
+        var onRepostPromptCancelledCalled = false
+        var request: PromptRequest.Repost = mock()
+
+        mockSession.register(object : EngineSession.Observer {
+            override fun onPromptRequest(promptRequest: PromptRequest) {
+                request = promptRequest as PromptRequest.Repost
+            }
+
+            override fun onRepostPromptCancelled() {
+                onRepostPromptCancelledCalled = true
+            }
+        })
+        val prompt = mock<GeckoRepostPrompt>()
+        doReturn(false).`when`(prompt).isComplete
+
+        GeckoPromptDelegate(mockSession).onRepostConfirmPrompt(mock(), prompt)
+        request.onDismiss()
+
+        assertTrue(onRepostPromptCancelledCalled)
+    }
+
+    @Test
     fun `dismissSafely only dismiss if the prompt is NOT already dismissed`() {
         val prompt = spy(GeckoAlertPrompt())
         val geckoResult = mock<GeckoResult<GeckoSession.PromptDelegate.PromptResponse>>()
@@ -1175,6 +1267,8 @@ class GeckoPromptDelegateTest {
     ) : GeckoSession.PromptDelegate.AutocompleteRequest<Autocomplete.LoginSaveOption>(login)
 
     class GeckoAuthOptions : GeckoSession.PromptDelegate.AuthPrompt.AuthOptions()
+
+    class GeckoRepostPrompt : GeckoSession.PromptDelegate.RepostConfirmPrompt()
 
     private fun GeckoSession.PromptDelegate.BasePrompt.getGeckoResult(): GeckoBundle {
         val javaClass = GeckoSession.PromptDelegate.BasePrompt::class.java